+++ src/main/java/com/example/gabojago_server/dto/response/article/accompany/AccompanyResponseDto.java
@@ -1,15 +1,19 @@
-package com.example.gabojago_server.dto.response.article;
+package com.example.gabojago_server.dto.response.article.accompany;
 
 import com.example.gabojago_server.model.article.AccompanyArticle;
 import com.example.gabojago_server.model.member.Member;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.format.DateTimeFormatter;
 
 @Getter
+@AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class PageAccompanyResponseDto {
+public class AccompanyResponseDto {
     private Long id;
 
     private Member writer;
@@ -28,8 +32,10 @@ public class PageAccompanyResponseDto {
 
     private int recruitMember;
 
-    public static PageAccompanyResponseDto of(AccompanyArticle article) {
-        return PageAccompanyResponseDto.builder()
+    private boolean isWritten;
+
+    public static AccompanyResponseDto of(AccompanyArticle article, boolean isWritten) {
+        return AccompanyResponseDto.builder()
                 .id(article.getId())
                 .writer(article.getWriter())
                 .title(article.getTitle())
@@ -39,6 +45,8 @@ public class PageAccompanyResponseDto {
                 .startDate(article.getStartDate().format(DateTimeFormatter.ofPattern("yyyy-MM-dd")))
                 .endDate(article.getEndDate().format(DateTimeFormatter.ofPattern("yyyy-MM-dd")))
                 .recruitMember(article.getRecruitNumber())
+                .isWritten(isWritten)
                 .build();
+
     }
 }
