+++ src/main/java/com/example/gabojago_server/repository/article/accompany/AccompanyArticleRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.example.gabojago_server.repository.article.accompany;
 
-import com.example.gabojago_server.dto.response.article.PageAccompanyResponseDto;
+import com.example.gabojago_server.dto.response.article.accompany.PageAccompanyResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
