+++ src/test/java/com/example/gabojago_server/service/article/AccompanyServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.gabojago_server.service.article;
 
 import com.example.gabojago_server.config.JpaConfig;
-import com.example.gabojago_server.dto.response.article.AccompanyResponseDto;
+import com.example.gabojago_server.dto.response.article.accompany.AccompanyResponseDto;
 import com.example.gabojago_server.model.article.AccompanyArticle;
 import com.example.gabojago_server.model.member.Member;
 import com.example.gabojago_server.repository.article.accompany.AccompanyArticleRepository;
