+++ src/test/java/com/example/gabojago_server/service/article/QnaServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.gabojago_server.service.article;
 
 import com.example.gabojago_server.config.JpaConfig;
-import com.example.gabojago_server.dto.response.article.QnaResponseDto;
+import com.example.gabojago_server.dto.response.article.qna.QnaResponseDto;
 import com.example.gabojago_server.model.member.Member;
 import com.example.gabojago_server.repository.article.Qna.QnaArticleRepository;
 import com.example.gabojago_server.repository.member.MemberRepository;
